mavlink: move COLLISION to streams hpp
